server/requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
Django==3.2.6
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
pika==1.2.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
service-identity==21.1.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
